log_getter added and tested

+++ test/log_getter_test.rb
@@ -1,6 +1,7 @@
 require "config"
 require "test/unit"
-require 'digest/md5'
+require "digest/md5"
+require "lib/log_getter"
 
 class LogGetterTest < Test::Unit::TestCase
   @@remote_file = "/mnt/src/parsa/test/data/access.log.gz"
@@ -8,20 +9,24 @@ class LogGetterTest < Test::Unit::TestCase
   
   def setup
     puts "deleting #{@@local_file}"
-    File.delete(@@local_file)
+    File.delete(@@local_file) if File.exists?(@@local_file)
   end
   
   def test_get_file
     remote_file_md5_sum = "d472821707a0cc2d25e884b9766ebe8f"
     local_file_md5_sum = ""
     
-    Net::SFTP.start('ubuntu', 'oleg', :password => 'oleg') do |sftp|
-      sftp.download!(@@remote_file, @@local_file)
-    end
+    data = LogGetterData.new
+    data.host = "ubuntu"
+    data.username = "oleg"
+    data.password = "oleg"
+    data.remote_file = @@remote_file
+    data.local_file = @@local_file
+
+    log_getter = LogGetter.new
+    log_getter.get_log(data)
     
     local_file_md5_sum = Digest::MD5.hexdigest(File.read(@@local_file))
     assert_equal(local_file_md5_sum, remote_file_md5_sum)
   end
 end
-
-
